# Refactor UpgradeSdkEnumConverter to cache the result instead of using reflection each time it reads or writes JSON