--- a/resources/sdk/pureclouddotnet/templates/AssemblyInfo.mustache
+++ b/resources/sdk/pureclouddotnet/templates/AssemblyInfo.mustache
@@ -1,4 +1,5 @@
 using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
@@ -12,6 +13,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("{{packageCopyright}}")]
 [assembly: AssemblyTrademark("{{packageTrademark}}")]
 [assembly: AssemblyCulture("{{packageCulture}}")]
+[assembly: InternalsVisibleTo("{{packageName}}.Tests")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from