--- a/resources/sdk/pureclouddotnet/templates/test-csproj.mustache
+++ b/resources/sdk/pureclouddotnet/templates/test-csproj.mustache
@@ -42,6 +42,7 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -81,6 +82,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SdkTests.cs" />
     <Compile Include="ApiClientTests.cs" />
+    <Compile Include="UpgradeSdkEnumConverterTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />